+++ xs/sources/driver/project_driver.h
@@ -6,9 +6,11 @@
 #ifndef XS_DRIVER_PROJECT_DRIVER_H
 #define XS_DRIVER_PROJECT_DRIVER_H
 
+#include "options.h"
+
 #include <stddef.h>
 
-bool xs_driver_resolve_kotlin_project(char ***paths, size_t *path_count);
+bool xs_driver_resolve_kotlin_project(char ***paths, size_t *path_count, XsCompilerSettings *settings);
 void xs_driver_free_project_paths(char **paths, size_t path_count);
 
 #endif